requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Babel==3.1.0
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
stripe==6.6.0
openai==0.28.1
python-dotenv==1.0.0
email-validator==2.0.0
Pillow==10.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
